+++ warthog/transport.py
@@ -30,10 +30,14 @@ from requests.packages.urllib3.exceptions import InsecureRequestWarning
 # Default to using the SSL/TLS version that the A10 requires instead of
 # the default that the requests/urllib3 library picks. Or, maybe the A10
 # just doesn't allow the client to negotiate. Either way, we use TLSv1.
+# pylint: disable=no-member
 DEFAULT_SSL_VERSION = ssl.PROTOCOL_TLSv1
 
+# Default to verifying SSL/TLS certs because "safe by default" is a good idea.
+DEFAULT_CERT_VERIFY = True
 
-def get_transport_factory(verify=True, ssl_version=DEFAULT_SSL_VERSION):
+
+def get_transport_factory(verify=None, ssl_version=None):
     """Get a new callable that returns :class:`requests.Session` instances that
     have been configured according to the given parameters.
 
@@ -44,27 +48,34 @@ def get_transport_factory(verify=True, ssl_version=DEFAULT_SSL_VERSION):
     you have special requirements such as needing to bypass HTTPS certificate validation
     because you use a self signed certificate.
 
-    :param bool verify: Should SSL certificates by verified when connecting
+    .. versionchanged:: 0.10.0
+        Using the requests/urllib3 default is no longer an option. Passing a ``None`` value
+        for ``ssl_version`` will result in using the Warthog default (TLS v1).
+
+    :param bool|None verify: Should SSL certificates by verified when connecting
         over HTTPS? Default is ``True``. If you have chosen not to verify certificates
         warnings about this emitted by the requests library will be suppressed.
-    :param int ssl_version: Explicit version of SSL to use for HTTPS connections
+    :param int|None ssl_version: Explicit version of SSL to use for HTTPS connections
         to an A10 load balancer. The version is a constant as specified by the
-        :mod:`ssl` module. The default is TLSv1. If you don't wish to use a specific
-        version and instead rely on the default for the requests / urllib3 module,
-        pass ``ssl_version=None``.
+        :mod:`ssl` module. The default is TLSv1.
     :return: A callable to return new configured session instances for making HTTP(S)
         requests
     :rtype: callable
     """
+    # Using `None` here to represent "not specified" so we don't have to litter the
+    # whole lib with references to the default values we've specified here. Callers
+    # just pass `None` and we'll pick the default here.
+    verify = verify if verify is not None else DEFAULT_CERT_VERIFY
+    ssl_version = ssl_version if ssl_version is not None else DEFAULT_SSL_VERSION
+
     # pylint: disable=missing-docstring
     def factory():
         transport = requests.Session()
+        transport.mount('https://', VersionedSSLAdapter(ssl_version))
 
         if not verify:
             transport.verify = False
 
-        if ssl_version is not None:
-            transport.mount('https://', VersionedSSLAdapter(ssl_version))
         return transport
 
     # Make sure that we suppress warnings about invalid certs since the user